structure custom data types
