+++ src/ui/activate.c
@@ -4,5 +4,5 @@
 void activate(GtkApplication *app, gpointer data) {
   GtkWidget *window = create_main_window(app);
   gtk_window_present(GTK_WINDOW(window));
-  g_print("Window is presented on the screen");
+  g_print("Window is presented on the screen\n");
 }
